Working on putting together a menu building function that will allow the ability to create menus on the fly.

+++ projects/manager/src/app/pages/niche-shack-editor/product-editor/product-content/product-content.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit, ViewChild, ViewChildren, ElementRef, QueryList } from '@angular/core';
 import { FormService } from 'projects/manager/src/app/services/form.service';
 import { ProductContent } from 'projects/manager/src/app/classes/product-content';
+import { MenuService } from 'projects/manager/src/app/services/menu.service';
 
 @Component({
   selector: 'product-content',
@@ -8,7 +9,7 @@ import { ProductContent } from 'projects/manager/src/app/classes/product-content
   styleUrls: ['./product-content.component.scss']
 })
 export class ProductContentComponent implements OnInit {
-  constructor(public _FormService: FormService) {}
+  constructor(public _FormService: FormService, public menuService: MenuService) {}
   public productContent: ProductContent = new ProductContent();
   public selectedColumnIndex: number = null;
   public selectedRowIndex: number = null;
@@ -71,6 +72,8 @@ export class ProductContentComponent implements OnInit {
       this.unsetEventListeners();
     }
   };
+
+  
 
 
   // -----------------------------( SET EVENT LISTENERS )------------------------------ \\
@@ -182,21 +185,73 @@ export class ProductContentComponent implements OnInit {
     this.productContent.selectedPricePointIndex = null;
   }
 
+  subMenu(width: number, name: string, ...options: any) {
+    return {type: "sub menu", width: width, name: name, options: options}
+  }
+
+  option(name: string, shortcutKeys?: string) {
+    return {type: "option", name: name, shortcutKeys: shortcutKeys}
+  }
+
+  divider() {
+    return {type: "divider"}
+  }
 
   // -----------------------------( SET CONTEXT MENU )------------------------------ \\
   setContextMenu(e: MouseEvent) {
     if (e.which == 3) {
-      // Define which selector we are clicking on
-      let selectorType: string = this.selectedColumnIndex != null ? "Price Point" : this.selectedRowIndex != null ? "Item" : null;
 
-      // Then name the menu options accordingly
-      this.showContextMenu = true;
-      this.direction1 = selectorType == "Price Point" ? "Left" : "Above";
-      this.direction2 = selectorType == "Price Point" ? "Right" : "Below";
-      this.contextMenuLeft = ((e.clientX - this.contextMenuOffset.nativeElement.getBoundingClientRect().x) + 25);
-      this.contextMenuTop = (e.clientY - this.contextMenuOffset.nativeElement.getBoundingClientRect().y) - (selectorType == "Price Point" ? 20 : 230);
-      this.pasteEnabled = (selectorType == 'Price Point' && this.copied == 'Price Point') || (selectorType == 'Item' && this.copied == 'Item') ? true : false;
-      this.deleteEnabled = (selectorType == 'Price Point' && this.productContent.items[0].pricePointOptions.length > 1) || (selectorType == 'Item' && this.productContent.items.length > 1) ? true : false;
+
+
+      this.menuService.buildMenu(100, 100, 300, this.option("Alita", "Ctrl+A"), 
+                                                this.divider(),
+                                                this.subMenu(100, "Sub Menu 1", this.option("Option A", "Ctrl+Shift+A"),
+                                                                                this.option("Option B", "Ctrl+Shift+B"),
+                                                                                this.subMenu(200, "Sub Menu 2", this.option("Option 2A", "Ctrl+Shift+2A"),
+                                                                                                                this.option("Option 2B", "Ctrl+Shift+2B"),
+                                                                                                                this.divider(),
+                                                                                                                this.option("Option 2C", "Ctrl+Shift+2C")),
+                                                                                this.option("Option C", "Ctrl+Shift+C"),
+                                                                                this.option("Option D", "Ctrl+Shift+D"),
+                                                                                this.divider(),
+                                                                                this.subMenu(300, "Sub Menu 3", this.option("Option 3A", "Ctrl+Shift+3A"),
+                                                                                                                this.option("Option 3B", "Ctrl+Shift+3B"),
+                                                                                                                this.divider(),
+
+
+
+                                                                                                                this.subMenu(400, "Sub Menu 4", this.option("Option 4A", "Ctrl+Shift+4A"),
+                                                                                                                                                this.option("Option 4B", "Ctrl+Shift+4B")),
+                                                                                                                this.option("Option 3C", "Ctrl+Shift+3C")),
+                                                                                this.option("Option E", "Ctrl+Shift+E"),
+                                                                                this.option("Option F", "Ctrl+Shift+F"),                           
+                                                                                ),
+                                                this.option("Battle", "Ctrl+B"),
+                                                this.option("Angel")
+      
+      );
+
+
+
+
+      
+
+
+
+
+
+      
+      // // Define which selector we are clicking on
+      // let selectorType: string = this.selectedColumnIndex != null ? "Price Point" : this.selectedRowIndex != null ? "Item" : null;
+
+      // // Then name the menu options accordingly
+      // this.showContextMenu = true;
+      // this.direction1 = selectorType == "Price Point" ? "Left" : "Above";
+      // this.direction2 = selectorType == "Price Point" ? "Right" : "Below";
+      // this.contextMenuLeft = ((e.clientX - this.contextMenuOffset.nativeElement.getBoundingClientRect().x) + 25);
+      // this.contextMenuTop = (e.clientY - this.contextMenuOffset.nativeElement.getBoundingClientRect().y) - (selectorType == "Price Point" ? 20 : 230);
+      // this.pasteEnabled = (selectorType == 'Price Point' && this.copied == 'Price Point') || (selectorType == 'Item' && this.copied == 'Item') ? true : false;
+      // this.deleteEnabled = (selectorType == 'Price Point' && this.productContent.items[0].pricePointOptions.length > 1) || (selectorType == 'Item' && this.productContent.items.length > 1) ? true : false;
     }
   }
 
@@ -346,23 +401,27 @@ export class ProductContentComponent implements OnInit {
 
   // -----------------------------( DELETE )------------------------------ \\
   delete() {
-    // If we're deleting a Price Point
-    if (this.selectedColumnIndex != null && this.productContent.items[0].pricePointOptions.length > 1) {
 
-      // Delete the Price Point at the specified index
-      this.productContent.pricePoints.splice(this.selectedColumnIndex, 1);
-      // And delete all the price point options from the deleted Price Point as well
-      for (let i = 0; i < this.productContent.items.length; i++) {
-        this.productContent.items[i].pricePointOptions.splice(this.selectedColumnIndex, 1);
+    if(this.deleteEnabled) {
+
+      // If we're deleting a Price Point
+      if (this.selectedColumnIndex != null && this.productContent.items[0].pricePointOptions.length > 1) {
+
+        // Delete the Price Point at the specified index
+        this.productContent.pricePoints.splice(this.selectedColumnIndex, 1);
+        // And delete all the price point options from the deleted Price Point as well
+        for (let i = 0; i < this.productContent.items.length; i++) {
+          this.productContent.items[i].pricePointOptions.splice(this.selectedColumnIndex, 1);
+        }
+
+        // If we're deleting an Item
+      } else if (this.selectedRowIndex != null && this.productContent.items.length > 1) {
+
+        // Delete the item at the specified index
+        this.productContent.items.splice(this.selectedRowIndex, 1);
       }
-
-      // If we're deleting an Item
-    } else if (this.selectedRowIndex != null && this.productContent.items.length > 1) {
-
-      // Delete the item at the specified index
-      this.productContent.items.splice(this.selectedRowIndex, 1);
+      this.showContextMenu = false;
     }
-    this.showContextMenu = false;
   }
 
 
@@ -575,8 +634,10 @@ export class ProductContentComponent implements OnInit {
 
   // -----------------------------( CUT )------------------------------ \\
   cut() {
-    this.copy();
-    this.delete();
+    if(this.deleteEnabled) {
+      this.copy();
+      this.delete();
+    }
   }
 
 
@@ -629,38 +690,42 @@ export class ProductContentComponent implements OnInit {
 
   // -----------------------------( PASTE )------------------------------ \\
   paste() {
-    // If we're pasting a price point
-    if (this.selectedColumnIndex != null && this.copied == "Price Point") {
-      // If the copied price point length is greater than the current price point length, then use the current price point length, otherwise, use the copied price point length
-      let pricePointOptionsLength = this.pricePointClipboard.pricePointOptions.length > this.productContent.items.length ? this.productContent.items.length : this.pricePointClipboard.pricePointOptions.length;
 
-      // Paste the price point value
-      this.productContent.pricePoints[this.selectedColumnIndex].textBefore = this.pricePointClipboard.pricePoint.textBefore;
-      this.productContent.pricePoints[this.selectedColumnIndex].wholeNumber = this.pricePointClipboard.pricePoint.wholeNumber;
-      this.productContent.pricePoints[this.selectedColumnIndex].decimal = this.pricePointClipboard.pricePoint.decimal;
-      this.productContent.pricePoints[this.selectedColumnIndex].textAfter = this.pricePointClipboard.pricePoint.textAfter;
+      if(this.pasteEnabled) {
 
-      // Paste all the price point option values
-      for (let i = 0; i < pricePointOptionsLength; i++) {
-        this.productContent.items[i].pricePointOptions[this.selectedColumnIndex] = this.pricePointClipboard.pricePointOptions[i];
+      // If we're pasting a price point
+      if (this.selectedColumnIndex != null && this.copied == "Price Point") {
+        // If the copied price point length is greater than the current price point length, then use the current price point length, otherwise, use the copied price point length
+        let pricePointOptionsLength = this.pricePointClipboard.pricePointOptions.length > this.productContent.items.length ? this.productContent.items.length : this.pricePointClipboard.pricePointOptions.length;
+
+        // Paste the price point value
+        this.productContent.pricePoints[this.selectedColumnIndex].textBefore = this.pricePointClipboard.pricePoint.textBefore;
+        this.productContent.pricePoints[this.selectedColumnIndex].wholeNumber = this.pricePointClipboard.pricePoint.wholeNumber;
+        this.productContent.pricePoints[this.selectedColumnIndex].decimal = this.pricePointClipboard.pricePoint.decimal;
+        this.productContent.pricePoints[this.selectedColumnIndex].textAfter = this.pricePointClipboard.pricePoint.textAfter;
+
+        // Paste all the price point option values
+        for (let i = 0; i < pricePointOptionsLength; i++) {
+          this.productContent.items[i].pricePointOptions[this.selectedColumnIndex] = this.pricePointClipboard.pricePointOptions[i];
+        }
+
+        // If we're pasting an Item
+      } else if (this.selectedRowIndex != null && this.copied == "Item") {
+
+        // If the copied price point length is greater than the current price point length, then use the current price point length, otherwise, use the copied price point length
+        let pricePointOptionsLength = this.itemClipboard.pricePointOptions.length > this.productContent.items[this.selectedRowIndex].pricePointOptions.length ? this.productContent.items[this.selectedRowIndex].pricePointOptions.length : this.itemClipboard.pricePointOptions.length;
+
+        // Paste the item's values
+        this.productContent.items[this.selectedRowIndex].type = this.itemClipboard.type;
+        this.productContent.items[this.selectedRowIndex].description = this.itemClipboard.description;
+        this.productContent.items[this.selectedRowIndex].showPlaceholder = this.itemClipboard.showPlaceholder;
+        // Paste all the price point option values
+        for (let i = 0; i < pricePointOptionsLength; i++) {
+          this.productContent.items[this.selectedRowIndex].pricePointOptions[i] = this.itemClipboard.pricePointOptions[i];
+        }
       }
-
-      // If we're pasting an Item
-    } else if (this.selectedRowIndex != null && this.copied == "Item") {
-
-      // If the copied price point length is greater than the current price point length, then use the current price point length, otherwise, use the copied price point length
-      let pricePointOptionsLength = this.itemClipboard.pricePointOptions.length > this.productContent.items[this.selectedRowIndex].pricePointOptions.length ? this.productContent.items[this.selectedRowIndex].pricePointOptions.length : this.itemClipboard.pricePointOptions.length;
-
-      // Paste the item's values
-      this.productContent.items[this.selectedRowIndex].type = this.itemClipboard.type;
-      this.productContent.items[this.selectedRowIndex].description = this.itemClipboard.description;
-      this.productContent.items[this.selectedRowIndex].showPlaceholder = this.itemClipboard.showPlaceholder;
-      // Paste all the price point option values
-      for (let i = 0; i < pricePointOptionsLength; i++) {
-        this.productContent.items[this.selectedRowIndex].pricePointOptions[i] = this.itemClipboard.pricePointOptions[i];
-      }
+      this.showContextMenu = false;
     }
-    this.showContextMenu = false;
   }
